Revert "Add light indicator for bursts to BsCostModel, #1336 (#1337)"
This reverts commit 3b1a363cecc7a179a26519f4c3f3964c43183c2b.
